+++ app/javascript/packs/application.js
@@ -291,7 +291,6 @@ window.Application = defaultExport.Application;
 
 $(() => {
   Application.enableDateTimePicker();
-  Application.enableOrderTypePicker(true);
   Application.enableOrderAttributePicker(false);
 });